add zoomZak and zoomSDKJWT as field resolver
